simplify reporter coordinator
