+++ src/commands/test/testing_summary.py
@@ -1,0 +1,135 @@
+from collections import defaultdict
+from logging import Logger
+from pathlib import Path
+from typing import Dict, List
+
+from src.commands.test.cases import BrokenTest, CaseResult, FailedCase, PassedCase
+from src.utils.log_color_provider import log_color_provider
+
+
+class TestingSummary:
+    def __init__(self, case_results: List[CaseResult]) -> None:
+        self.case_results = []
+        self.test_files: Dict[Path, List[CaseResult]] = defaultdict(list)
+        self.passed: List[PassedCase] = []
+        self.failed: List[FailedCase] = []
+        self.broken: List[BrokenTest] = []
+        self.extend(case_results)
+
+    def extend(self, case_results: List[CaseResult]):
+        self.case_results += case_results
+        for case_result in case_results:
+            self.test_files[case_result.file_path].append(case_result)
+
+            if isinstance(case_result, PassedCase):
+                self.passed.append(case_result)
+            if isinstance(case_result, FailedCase):
+                self.failed.append(case_result)
+            if isinstance(case_result, BrokenTest):
+                self.broken.append(case_result)
+
+    def log(
+        self,
+        logger: Logger,
+        collected_test_cases_count: int,
+        collected_test_files_count: int,
+    ):
+        logger.info(
+            log_color_provider.bold("Test suits: ")
+            + self._get_test_suits_summary(collected_test_files_count)
+        )
+        logger.info(
+            log_color_provider.bold("Tests:      ")
+            + self._get_test_cases_summary(collected_test_cases_count)
+        )
+
+    def _get_test_cases_summary(self, collected_test_cases_count: int) -> str:
+        failed_test_cases_count = len(self.failed)
+        passed_test_cases_count = len(self.passed)
+
+        return ", ".join(
+            self._get_preprocessed_core_testing_summary(
+                failed_count=failed_test_cases_count,
+                passed_count=passed_test_cases_count,
+                total_count=collected_test_cases_count,
+            )
+        )
+
+    def _get_test_suits_summary(self, collected_test_files_count: int) -> str:
+        passed_test_suits_count = 0
+        failed_test_suits_count = 0
+        broken_test_suits_count = 0
+        total_test_suits_count = len(self.test_files)
+        for suit_case_results in self.test_files.values():
+            partial_summary = TestingSummary(suit_case_results)
+
+            if len(partial_summary.broken) > 0:
+                broken_test_suits_count += 1
+                continue
+
+            if len(partial_summary.failed) > 0:
+                failed_test_suits_count += 1
+                continue
+
+            if len(partial_summary.passed) > 0:
+                passed_test_suits_count += 1
+
+        test_suits_result: List[str] = []
+
+        if broken_test_suits_count > 0:
+            test_suits_result.append(
+                log_color_provider.colorize("RED", f"{broken_test_suits_count} broken")
+            )
+        if failed_test_suits_count > 0:
+            test_suits_result.append(
+                log_color_provider.colorize("RED", f"{failed_test_suits_count} failed")
+            )
+        if passed_test_suits_count > 0:
+            test_suits_result.append(
+                log_color_provider.colorize(
+                    "GREEN", f"{passed_test_suits_count} passed"
+                )
+            )
+        if total_test_suits_count > 0:
+            test_suits_result.append(f"{total_test_suits_count} total")
+
+        return ", ".join(
+            self._get_preprocessed_core_testing_summary(
+                broken_count=broken_test_suits_count,
+                failed_count=failed_test_suits_count,
+                passed_count=passed_test_suits_count,
+                total_count=collected_test_files_count,
+            )
+        )
+
+    # pylint: disable=no-self-use
+    def _get_preprocessed_core_testing_summary(
+        self,
+        broken_count: int = 0,
+        failed_count: int = 0,
+        passed_count: int = 0,
+        total_count: int = 0,
+    ) -> List[str]:
+        skipped_count = total_count - (broken_count + failed_count + passed_count)
+        test_suits_result: List[str] = []
+
+        if broken_count > 0:
+            test_suits_result.append(
+                log_color_provider.colorize("RED", f"{broken_count} broken")
+            )
+        if failed_count > 0:
+            test_suits_result.append(
+                log_color_provider.colorize("RED", f"{failed_count} failed")
+            )
+        if skipped_count > 0:
+            test_suits_result.append(
+                log_color_provider.colorize("YELLOW", f"{skipped_count} skipped")
+            )
+        if passed_count > 0:
+            test_suits_result.append(
+                log_color_provider.colorize("GREEN", f"{passed_count} passed")
+            )
+        if total_count > 0:
+            test_suits_result.append(f"{total_count} total")
+
+        return test_suits_result
